fix: 경로 수정
- 파이프라인 코드 제작과정 -> pipelines
- 최종모델코드는 모델링4차_분류모델병합
