--- 04_models_1/modules/feature_selector.py
@@ -1,153 +1,0 @@
-# 선별된 피처 리스트 정리
-
-import numpy as np
-import pandas as pd
-
-# 나눗셈 함수
-def safe_div(a, b):
-    return np.where(b == 0, 0, a / b)
-
-# CD 파생변수 생성
-def generate_cd_derived_features(df):
-    """
-    CD 관련 파생변수 생성
-    """
-    df = df.copy()
-
-    # 파생 변수 생성
-    df['오프라인_소비급등비율'] = safe_div(df['이용금액_오프라인_B0M'], df['이용금액_오프라인_R6M'] / 6)
-    df['일시불_최근성지표'] = safe_div(df['이용금액_일시불_B0M'], df['최대이용금액_일시불_R12M'] + 1)
-    df['소비_평균대최대비율'] = safe_div((df['이용금액_R3M_신용'] + 1) / 3, df['최대이용금액_일시불_R12M'] + 1)
-    df['청구_최근성지표'] = safe_div(df['정상청구원금_B0M'], df['정상청구원금_B5M'] + 1)
-    df['소진율_차이'] = df['잔액_신판평균한도소진율_r6m'] - df['잔액_신판ca평균한도소진율_r6m']
-    df['카드집중도'] = safe_div(df['_1순위카드이용금액'], df['이용금액_R3M_신용'] + df['이용금액_R3M_신용체크'] + 1)
-    df['입금청구비율'] = safe_div(df['정상입금원금_B5M'], df['정상청구원금_B5M'] + 1)
-    df['일시불_금액비율'] = safe_div(df['이용금액_일시불_B0M'], df['평잔_일시불_6M'] + 1)
-    df['할인전이자율_평균'] = (df['RV일시불이자율_할인전'] + df['CA이자율_할인전']) / 2
-
-    return df
-
-ab_feat = [
-    '할부금액_3M_R12M', '이용금액_할부_무이자_R12M', '이용건수_할부_무이자_R12M', '정상입금원금_B0M',
-    '이용금액_오프라인_R6M', '정상청구원금_B0M', '이용금액_오프라인_R3M', '_1순위카드이용금액', '평잔_일시불_해외_6M',
-    '승인거절건수_입력오류_R3M', '청구금액_R3M', '이용금액_일시불_R12M', '이용금액_할부_무이자_R3M',
-    '이용금액_할부_무이자_R6M', '정상입금원금_B5M', '이용금액_할부_R12M', '마일_적립포인트_R3M',
-    '정상청구원금_B2M', '포인트_마일리지_환산_B0M', '청구금액_B0', '정상입금원금_B2M', '할부건수_무이자_3M_R12M',
-    '정상청구원금_B5M', '청구금액_R6M', '여유_숙박이용금액', '최대이용금액_일시불_R12M',
-    '_1순위업종_이용금액', '잔액_할부_B0M', '할부금액_무이자_3M_R12M','잔액_할부_무이자_B0M'
-    ]
-
-cd_baseline = [
-    '이용금액_일시불_R3M', '이용금액_R3M_신용체크', '정상입금원금_B0M', '이용금액_오프라인_R6M',
-    '정상청구원금_B0M', '이용건수_신용_R6M', '이용금액_오프라인_R3M', '이용건수_일시불_R12M',
-    '_1순위카드이용금액', '이용금액_오프라인_B0M', '청구금액_R3M', '이용금액_일시불_R12M',
-    '이용금액_R3M_신용', '정상입금원금_B5M', '이용건수_오프라인_B0M', '정상청구원금_B2M',
-    '이용건수_신판_R12M', '청구금액_B0', '정상입금원금_B2M', '이용금액_일시불_B0M',
-    '정상청구원금_B5M', '청구금액_R6M', '최대이용금액_일시불_R12M', '이용가맹점수',
-    '이용건수_신용_R12M', '이용금액_일시불_R6M', '잔액_신판평균한도소진율_r6m',
-    '이용금액대', '평잔_일시불_6M', '잔액_신판ca평균한도소진율_r6m',
-    'CA이자율_할인전', 'RV일시불이자율_할인전',
-
-    '평잔_카드론_3M', '평잔_카드론_6M', '잔액_카드론_B1M', '잔액_카드론_B2M',
-    '잔액_카드론_B3M', '잔액_카드론_B0M', '잔액_카드론_B4M', '잔액_카드론_B5M',
-    '카드론이용금액_누적', '최종카드론_대출금액',
-    '잔액_신판평균한도소진율_r3m', '잔액_신판최대한도소진율_r3m', '잔액_신판최대한도소진율_r6m',
-
-    '평잔_RV일시불_3M', '평잔_RV일시불_6M', 'RV_평균잔액_R3M',
-    '잔액_할부_B0M', '잔액_할부_B1M', '잔액_리볼빙일시불이월_B0M',
-    '평잔_할부_3M', '잔액_할부_B2M', '평잔_할부_6M',
-    '잔액_일시불_B1M', '잔액_일시불_B2M', '잔액_할부_무이자_B0M',
-    '잔액_일시불_B0M', '평잔_일시불_3M'
-]
-
-cd_final_features = [
-    '정상청구원금_B5M', '이용금액_R3M_신용체크', '정상입금원금_B5M', 
-    '최대이용금액_일시불_R12M', '이용가맹점수', '이용금액_오프라인_R6M', 
-    '평잔_일시불_6M', '잔액_신판ca평균한도소진율_r6m', 'RV일시불이자율_할인전', '오프라인_소비급등비율', '일시불_최근성지표', 
-    '소비_평균대최대비율', '청구_최근성지표', '소진율_차이', '카드집중도', '입금청구비율', '일시불_금액비율'
-    ]
-
-# 📌 PC1~PC5에서 반복적으로 중요한 변수:
-pca_cols = [
-    'CA이자율_할인전', 'CL이자율_할인전', 'RV_평균잔액_R3M', 'RV일시불이자율_할인전', 
-    'RV최소결제비율', 'RV현금서비스이자율_할인전', '방문월수_앱_R6M', '방문일수_앱_B0M', '방문일수_앱_R6M', '방문횟수_앱_B0M', '방문후경과월_앱_R6M', 
-    '이용금액_R3M_신용', '이용금액_R3M_신용체크', '이용금액_일시불_B0M', '이용금액대', '일시불ONLY전환가능여부', 
-    '잔액_리볼빙일시불이월_B0M', '잔액_일시불_B0M', '잔액_일시불_B1M', '잔액_일시불_B2M', '잔액_카드론_B0M', '잔액_카드론_B1M', '잔액_카드론_B2M', '잔액_카드론_B3M', '잔액_카드론_B4M', 
-    '잔액_카드론_B5M', '정상청구원금_B0M', '정상청구원금_B2M', '정상청구원금_B5M', '청구금액_B0', '청구금액_R3M', '청구금액_R6M', '최종카드론_대출금액', 
-    '카드론이용금액_누적', '평잔_RV일시불_3M', '평잔_RV일시불_6M', '평잔_일시불_3M', '평잔_일시불_6M', '평잔_카드론_3M', '평잔_카드론_6M', '평잔_할부_3M', 
-    '홈페이지_금융건수_R3M', '홈페이지_금융건수_R6M', '홈페이지_선결제건수_R3M', '홈페이지_선결제건수_R6M']
-
-top_cde = [
-    '이용금액_일시불_R3M', '이용금액_R3M_신용체크', '정상입금원금_B0M', '이용금액_오프라인_R6M', '정상청구원금_B0M', '이용건수_신용_R6M', '이용금액_오프라인_R3M','이용건수_일시불_R12M',
-    '_1순위카드이용금액', '이용금액_오프라인_B0M', '청구금액_R3M', '이용금액_일시불_R12M', '이용금액_R3M_신용',
-    '정상입금원금_B5M', '이용건수_오프라인_B0M', '정상청구원금_B2M', '이용건수_신판_R12M', '청구금액_B0',
-    '정상입금원금_B2M', '이용금액_일시불_B0M', '정상청구원금_B5M', '청구금액_R6M', '최대이용금액_일시불_R12M',
-    '이용가맹점수', '이용건수_신용_R12M', '이용금액_일시불_R6M'
-    ]
-
-selected_cols=(ab_feat + top_cde + pca_cols)
-selected_cols = list(dict.fromkeys(selected_cols))
-
-#baseline features
-baseline_feat = [
-    '남녀구분코드', '회원여부_이용가능_카드론', '입회경과개월수_신용', '회원여부_연체', '동의여부_한도증액안내',
-    '탈회횟수_누적', '탈회횟수_발급1년이내', '거주시도명', '유효카드수_신용체크', '이용카드수_신용체크',
-    '이용금액_R3M_신용체크', '보유여부_해외겸용_본인', '이용여부_3M_해외겸용_본인', '연회비발생카드수_B0M', '기본연회비_B0M',
-    '제휴연회비_B0M', '최초한도금액', '카드이용한도금액', 'CA한도금액', '일시상환론한도금액',
-    '월상환론한도금액', 'CA이자율_할인전', 'RV일시불이자율_할인전', 'RV현금서비스이자율_할인전', 'RV약정청구율',
-    'RV최소결제비율', '한도증액횟수_R12M', '한도증액금액_R12M', '상향가능CA한도금액', '월상환론상향가능한도금액',
-    '일시불ONLY전환가능여부', '특별한도보유여부_R3M', '연체감액여부_R3M', '이용건수_신용_B0M', '이용금액_일시불_B0M',
-    '이용후경과월_신용', '이용후경과월_일시불', '이용후경과월_체크', '이용금액_해외', '쇼핑_도소매_이용금액',
-    '쇼핑_마트_이용금액', '쇼핑_온라인_이용금액', '교통_주유이용금액', '교통_정비이용금액', '여유_숙박이용금액',
-    '할부건수_3M_R12M', '할부금액_3M_R12M', 'RP건수_B0M', 'RP건수_통신_B0M', 'RP건수_보험_B0M',
-    'RP건수_교통_B0M', 'RP후경과월', '카드론이용건수_누적', '카드론이용금액_누적', '최종카드론_대출이율',
-    '최종카드론_대출금액', '이용건수_온라인_R6M', '이용건수_페이_온라인_R6M', '이용건수_A페이_R6M', '이용건수_C페이_R6M',
-    '이용건수_D페이_R6M', '이용금액_선결제_R6M', '이용건수_선결제_R6M', '연체입금원금_B0M', '정상청구원금_B5M',
-    '선입금원금_B2M', '연속유실적개월수_기본_24M_카드', '승인거절건수_R3M', '승인거절건수_한도초과_R3M', '이용금액대',
-    '청구서발송여부_B0', '청구서발송여부_R3M', '청구서발송여부_R6M', '청구금액_B0', '청구금액_R3M',
-    '청구금액_R6M', '포인트_마일리지_건별_B0M', '포인트_포인트_건별_B0M', '포인트_마일리지_월적립_B0M', '포인트_포인트_월적립_B0M',
-    '마일_적립포인트_R3M', '마일_잔여포인트_B0M', '포인트_이용포인트_R12M', '할인건수_R3M', '할인금액_R3M',
-    '할인건수_B0M', '혜택수혜금액_R3M', '포인트_마일리지_환산_B0M', '잔액_일시불_B0M', '잔액_할부_B0M',
-    '연체잔액_B0M', '연체원금_최근', '매각잔액_B1M', 'RV_평균잔액_R3M', '연체잔액_RV일시불_B0M',
-    '연체잔액_CA_B0M', '평잔_일시불_3M', '평잔_6M', '평잔_카드론_3M', '평잔_CA_3M',
-    'IB문의건수_사용승인내역_R6M', 'IB문의건수_선결제_R6M', '당사멤버쉽_방문월수_R6M', '당사멤버쉽_방문횟수_B0M', '당사멤버쉽_방문횟수_R6M',
-    '방문월수_PC_R6M', '방문월수_앱_R6M', '방문일수_PC_B0M', '방문일수_PC_R6M', '방문일수_앱_B0M',
-    '방문일수_앱_R6M', '방문횟수_PC_B0M', '방문횟수_PC_R6M', '방문횟수_앱_B0M', '방문횟수_앱_R6M',
-    '방문후경과월_PC_R6M', '방문후경과월_앱_R6M', '불만제기후경과월_R12M', '상담건수_B0M', '상담건수_R6M',
-    '이용메뉴건수_ARS_B0M', '이용메뉴건수_IB_R6M', '인입일수_ARS_B0M', '인입월수_ARS_R6M', '인입일수_IB_B0M',
-    '인입월수_IB_R6M', '인입횟수_IB_R6M', '인입후경과월_ARS', '인입후경과월_IB_R6M', '홈페이지_금융건수_R3M',
-    '홈페이지_금융건수_R6M', '홈페이지_선결제건수_R3M', '홈페이지_선결제건수_R6M', '증감율_이용건수_할부_전월', '증감율_이용건수_CA_전월',
-    '증감율_이용건수_체크_전월', '증감율_이용금액_신용_전월', '증감율_이용금액_할부_전월', '증감율_이용금액_신판_전월', '증감율_이용금액_일시불_전월',
-    '증감율_이용건수_신판_분기', '증감율_이용건수_일시불_분기', '증감율_이용금액_신용_분기', '증감율_이용금액_체크_분기', '잔액_신판평균한도소진율_r6m',
-    '잔액_신판최대한도소진율_r6m', '잔액_신판ca평균한도소진율_r6m', '잔액_신판ca최대한도소진율_r6m', '변동률_RV일시불평잔', '변동률_할부평잔',
-    '변동률_RVCA평잔', '변동률_카드론평잔', '이용카드수_신용', '이용카드수_신용_가족', '이용카드수_체크',
-    '이용카드수_체크_가족', '이용금액_R3M_신용', '이용금액_R3M_신용_가족', '이용금액_R3M_체크', '이용금액_R3M_체크_가족',
-    'CL이자율_할인전', '상향가능한도금액', '카드이용한도금액_B1M', '카드이용한도금액_B2M', '이용후경과월_신판',
-    '이용후경과월_할부', '이용후경과월_할부_유이자', '이용후경과월_할부_무이자', '이용후경과월_부분무이자', '이용후경과월_CA',
-    '이용후경과월_카드론', '정상청구원금_B0M', '정상청구원금_B2M', '포인트_마일리지_건별_R3M', '포인트_포인트_건별_R3M',
-    '포인트_마일리지_월적립_R3M', '포인트_포인트_월적립_R3M', '포인트_적립포인트_R12M', '포인트_적립포인트_R3M', '포인트_이용포인트_R3M',
-    '포인트_잔여포인트_B0M', '잔액_현금서비스_B0M', '잔액_리볼빙일시불이월_B0M', '잔액_리볼빙CA이월_B0M', '잔액_카드론_B0M',
-    '잔액_현금서비스_B1M', '잔액_현금서비스_B2M', '잔액_카드론_B1M', '잔액_카드론_B2M', '잔액_카드론_B3M',
-    '잔액_카드론_B4M', '잔액_카드론_B5M', '잔액_할부_B1M', '잔액_할부_B2M', '잔액_일시불_B1M',
-    '잔액_일시불_B2M', '잔액_할부_유이자_B0M', '잔액_할부_무이자_B0M', '잔액_할부_해외_B0M', '평잔_3M',
-    '평잔_일시불_해외_3M', '평잔_RV일시불_3M', '평잔_RV일시불_해외_3M', '평잔_할부_3M', '평잔_할부_해외_3M',
-    '평잔_CA_해외_3M', '평잔_일시불_6M', '평잔_일시불_해외_6M', '평잔_RV일시불_6M', '평잔_RV일시불_해외_6M',
-    '평잔_할부_6M', '평잔_할부_해외_6M', '평잔_CA_6M', '평잔_CA_해외_6M', '평잔_카드론_6M',
-    '잔액_신판평균한도소진율_r3m', '잔액_신판최대한도소진율_r3m', '잔액_신판ca평균한도소진율_r3m', '잔액_신판ca최대한도소진율_r3m'
-]
-
-stage_feature_map = {
-    "base": baseline_feat,
-    "cd_base": cd_baseline,
-    "cd" : cd_baseline,
-    "ab" : ab_feat,
-    "e" : selected_cols
-}
-
-"""
-from feature_selector import stage_feature_map
-
-# '__' 단계의 피처만 불러오기
-selected_features = stage_feature_map['cd']
-입력 X : baseline (전체 피처)
-"""
